+++ mailgun.php
@@ -3,8 +3,8 @@
  * Plugin Name:  Mailgun
  * Plugin URI:   http://wordpress.org/extend/plugins/mailgun/
  * Description:  Mailgun integration for WordPress
- * Version:      1.9.5
- * Tested up to: 6.1
+ * Version:      1.9.6
+ * Tested up to: 6.3.1
  * Author:       Mailgun
  * Author URI:   http://www.mailgun.com/
  * License:      GPLv2 or later