+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Changelog
 =========
+1.9.6 (2023-09-10)
+- Added typehints to methods 
+- Added code for making plugin WP VIP GO compatible for the `From mail`
+- Added Toastr for sending nice message instead of using alerts
+
 1.9.5 (2023-06-20)
 - Fix bug with sending emails